// Frontend/src/app/interfaces/contrat.ts
export interface Contrat {
type_contrat: any
  moto_numero_serie: any
  chauffeur_username: any
  id:number,
  moto: number
  chauffeur:number
  montant_initial:string
  montant_journalier:string
  date_debut:string
  date_fin:string
  etat:string
  created_by: number
  modified_by: number
  created_at:string
  modified_at: string
  modified_by_name: any
  created_by_name: any
}